pkg/runtime: Various fixes to pass type-checking

+++ pkg/runtime/slices.go
@@ -35,7 +35,7 @@ func sliceappend(t unsafe.Pointer, a, b slice) slice {
 	}
 	end := uintptr(unsafe.Pointer(a.array))
 	end += uintptr(a.len) * slicetyp.elem.size
-	copylen := int(slicetyp.elem.size * uintptr(b.len))
+	copylen := slicetyp.elem.size * uintptr(b.len)
 	memmove(unsafe.Pointer(end), unsafe.Pointer(b.array), copylen)
 	a.len += b.len
 	return a
@@ -52,7 +52,7 @@ func slicecopy(t unsafe.Pointer, a, b slice) uint {
 	if b.len < n {
 		n = b.len
 	}
-	copylen := int(slicetyp.elem.size * uintptr(n))
+	copylen := slicetyp.elem.size * uintptr(n)
 	memmove(unsafe.Pointer(a.array), unsafe.Pointer(b.array), copylen)
 	return n
 }
@@ -61,19 +61,19 @@ func slicegrow(t *sliceType, a slice, newcap uint) slice {
 	mem := malloc(uintptr(t.elem.size * uintptr(newcap)))
 	if a.len > 0 {
 		size := uintptr(a.len) * t.elem.size
-		memcpy(mem, unsafe.Pointer(a.array), int(size))
+		memcpy(mem, unsafe.Pointer(a.array), size)
 	}
 	return slice{(*uint8)(mem), a.len, newcap}
 }
 
-func sliceslice(t unsafe.Pointer, a slice, low, high int32) slice {
+func sliceslice(t unsafe.Pointer, a slice, low, high int) slice {
 	if high == -1 {
-		high = a.len
+		high = int(a.len)
 	} else {
 		// TODO check upper bound
 	}
-	a.cap -= low
-	a.len = high - low
+	a.cap -= uint(low)
+	a.len = uint(high - low)
 	if low > 0 {
 		typ := (*type_)(t)
 		slicetyp := (*sliceType)(unsafe.Pointer(&typ.commonType))
